minor code updates
