+++ java/src/main/java/com/google/cloud/dataproc/templates/kafka/KafkaToPubSub.java
@@ -18,21 +18,18 @@ package com.google.cloud.dataproc.templates.kafka;
 import static com.google.cloud.dataproc.templates.util.TemplateConstants.*;
 
 import com.google.cloud.dataproc.templates.BaseTemplate;
-import com.google.cloud.spark.bigquery.repackaged.com.google.cloud.bigquery.*;
+import com.google.cloud.pubsub.v1.Publisher;
+import com.google.protobuf.ByteString;
+import com.google.pubsub.v1.PubsubMessage;
+import com.google.pubsub.v1.TopicName;
 import java.io.Serializable;
-import java.util.*;
 import java.util.Objects;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.apache.spark.SparkConf;
-import org.apache.spark.api.java.*;
-import org.apache.spark.api.java.function.*;
+import org.apache.spark.sql.Dataset;
+import org.apache.spark.sql.ForeachWriter;
+import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SparkSession;
-import org.apache.spark.streaming.Seconds;
-import org.apache.spark.streaming.api.java.*;
-import org.apache.spark.streaming.api.java.JavaStreamingContext;
-import org.apache.spark.streaming.kafka010.*;
+import org.apache.spark.sql.types.DataTypes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -70,8 +67,6 @@ public class KafkaToPubSub implements BaseTemplate, Serializable {
     }
 
     SparkSession spark = null;
-    SparkConf sparkConf = null;
-    JavaStreamingContext streamingContext = null;
     LOGGER.info(
         "Starting Kafka to PubSub spark job with following parameters:"
             + "1. {}:{}"
@@ -92,44 +87,83 @@ public class KafkaToPubSub implements BaseTemplate, Serializable {
 
     try {
       // Initialize the Spark session
-      // spark = SparkSession.builder().appName("Spark KafkaToPubSub Job").getOrCreate();
-      sparkConf = new SparkConf().setAppName("Spark KafkaToPubSub Job");
-      streamingContext = new JavaStreamingContext(sparkConf, Seconds.apply(15));
+      spark = SparkSession.builder().appName("Spark KafkaToPubSub Job").getOrCreate();
 
-      Map<String, Object> kafkaParams = new HashMap<>();
-      kafkaParams.put("bootstrap.servers", "10.0.0.20:9092");
-      kafkaParams.put("key.deserializer", StringDeserializer.class);
-      kafkaParams.put("value.deserializer", StringDeserializer.class);
-      // kafkaParams.put("group.id", "test-consumer-group"); // kafkaParams.put("group.id",
-      // "use_a_separate_group_id_for_each_stream");
-      kafkaParams.put("auto.offset.reset", "earliest");
-      kafkaParams.put("enable.auto.commit", false);
+      LOGGER.debug("added jars : {}", spark.sparkContext().addedJars().keys());
 
-      Collection<String> topics = Arrays.asList("test");
+      // Stream data from Kafka topic
+      Dataset<Row> inputData =
+          spark
+              .readStream()
+              .format("kafka")
+              .option("kafka.bootstrap.servers", kafkaBootstrapServers)
+              .option("subscribe", kafkaTopic)
+              .option("startingOffsets", kafkaStartingOffsets)
+              .option("failOnDataLoss", "false")
+              .load();
 
-      JavaInputDStream<ConsumerRecord<String, String>> stream =
-          KafkaUtils.createDirectStream(
-              streamingContext,
-              LocationStrategies.PreferConsistent(),
-              ConsumerStrategies.<String, String>Subscribe(topics, kafkaParams));
+      Dataset<Row> processedData =
+          inputData.withColumn("value", inputData.col("value").cast(DataTypes.StringType));
 
-      stream.foreachRDD(
-          rdd -> {
-            rdd.foreachPartition(
-                consumerRecords -> {
-                  System.out.println("consumerRecords: " + consumerRecords);
-                });
-          });
+      writeToPubSub(processedData, pubsubCheckpointLocation, kafkaAwaitTerminationTimeout);
 
-      streamingContext.start();
-      streamingContext.awaitTerminationOrTimeout(1200);
       LOGGER.info("KakfaToPubSub job completed.");
-      streamingContext.stop();
+      spark.stop();
     } catch (Throwable th) {
       LOGGER.error("Exception in KakfaToPubSub", th);
       if (Objects.nonNull(spark)) {
-        streamingContext.stop();
+        spark.stop();
       }
     }
+  }
+
+  // foreachRDD
+  // static void writeToPubSub(processedData,pubsubCheckpointLocation){
+  // }
+
+  // foreach
+  static void writeToPubSub(
+      Dataset<Row> processedData,
+      String pubsubCheckpointLocation,
+      Long kafkaAwaitTerminationTimeout)
+      throws Exception {
+    processedData
+        .select("value")
+        .writeStream()
+        .option("checkpointLocation", pubsubCheckpointLocation)
+        .foreach(
+            new ForeachWriter<Row>() {
+              Publisher publisher = null;
+              TopicName topicName = null;
+
+              @Override
+              public boolean open(long partitionId, long version) {
+                // Open connection
+                try {
+                  topicName = TopicName.of("yadavaja-sandbox", "kafkatopubsub");
+                  publisher = Publisher.newBuilder(topicName).build();
+                  return true;
+                } catch (Exception e) {
+                  // TODO: handle exception
+                  return true;
+                }
+              }
+
+              @Override
+              public void process(Row row) {
+                // Write string to connection
+                ByteString data = ByteString.copyFromUtf8(row.getString(0));
+                PubsubMessage pubsubMessage = PubsubMessage.newBuilder().setData(data).build();
+                publisher.publish(pubsubMessage);
+              }
+
+              @Override
+              public void close(Throwable errorOrNull) {
+                // Close the connection
+                // publisher.shutdown();
+              }
+            })
+        .start()
+        .awaitTermination(kafkaAwaitTerminationTimeout);
   }
 }
